Update ranges to reflect hardware PWM range, update default range
The hardware range checking should only check that we're in a hardware
range that makes sense. However, I understand why this change was
originally made (I assume out of annoyance with the default value), so
I've updated the default value and created a constructor that permits a
new starting value outside of the default.